requirements.txt
# Core ML / AI
langchain>=0.2.0
langgraph>=0.1.0
sentence-transformers>=2.7.0
faiss-cpu>=1.8.0
huggingface-hub>=0.23.0
transformers>=4.40.0
torch>=2.2.0

# API & Web
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
streamlit>=1.35.0

# MLOps
mlflow>=2.13.0

# Data & Utilities
pyyaml>=6.0
python-dotenv>=1.0.0
requests>=2.32.0
aiohttp>=3.9.0
beautifulsoup4>=4.12.0
tiktoken>=0.7.0

# Testing & Quality
pytest>=8.2.0
pytest-cov>=5.0.0
pytest-asyncio>=0.23.0
ruff>=0.4.0
httpx>=0.27.0

# Logging
rich>=13.7.0
structlog>=24.1.0
